Fix List Item width styling. (#271)
Recent change added 100% width to the list item contents, which extends it past its parent bounds.

Issue wasn't visible in Dewey because there was a top-level `box-sizing: border-box` applied to all elements in Dewey. Took that out so that components build in Dewey don't make assumptions about the base styling clients are using.