Optionally hide labels on the FileInput component
We would like to reuse the functionality of the `FileInput` component
without having the labels present.

This adds a new option `displayLabels` which can be set to `false` to
hide the labels on the `FileInput` component. The default value of
`displayLabels` is `true`. This allows all of the existing code using
the `FileInput` component to remain unchanged.