Clean up build directory before compilation
The first step performed by the build process is to wholesale copy the
entire source directory `src/` to the build directory `dist/`. At the
very end of the process, it deletes all of the compiled source files.

Here we remove the source files from the build directory before
compilation begins. This moves us a step closer to a world where we
explicitly move files from the source directory to the build directory
through a compilation or build step.