TypeScript build emits type declaration files
There is a chain of inheritance between these three TypeScript config
files; `tsconfig.docs.json` inherits from `tsconfig.build.json` which
in turn inherits from `tsconfig.json`.

Currently the `"allowJs": true` setting in the parent `tsconfig.json`
affects all three configuration files.

When we change the value in the middle of the inheritance chain, all
children receive the new value. To avoid this, we explicitly reset the
value of `"allowJs"` back to `true` in the child `tsconfig.docs.json`.

The end result is we set two new configuration values in the middle
`tsconfig.build.json` without altering the configuration of the child
`tsconfig.docs.json`.