Traverse type declaration subdirectories
Since a type declaration file exists now, TypeScript is using it in
consumers of the Clever Components library. When a component is not
referenced, it just uses type `any` -- which is fine for our use case.

However, when the reference is missing altogether, then TypeScript
throws an error. The folders referenced from the index file were not
getting added to the type declarations because the index files did not
have `.ts` file extensions.

Now, consumers of the Clever Components library should be aware of all
components exported from the top level `index.ts` file.