TextInputs do not display required error messages
We enhanced TextInputs to display an error message if they are
required and the user skips over the field without filling out a
value.

During these enhancements, we added this behavior in always.
Currently, if any TextInput is focused and then blurred without
entering a value, it will display an error message.

We only want this behavior if the TextInput is marked as required.