TextInputs announce errors and set aria flags
Aria attributes can be set on the client side or on the server side.
If they are set by the client, we need to notify screen readers that
something has been changed and needs to be announced. So we need an
`aria-live` tag on first render. If the contents of this `aria-live`
element change, the screen reader will be notified.